+++ react/training/src/app/screens/Login/index.js
@@ -1,16 +1,12 @@
-import React, { Fragment } from 'react';
+import React, { Component } from 'react';
 
-export const customInput = props => (
-  <Fragment>
-    <label> {props.label} </label>
-    <input {...props.input} type={props.type} />
-    {props.meta.error && <div style={{ color: 'red' }}> {props.meta.error} </div>}
-  </Fragment>
-);
+import LoginForm from './components/layout';
 
-export const required = value => (value ? undefined : 'Value is required');
+class Login extends Component {
+  submit = values => window.alert(JSON.stringify(values, null, 4));
+  render() {
+    return <LoginForm onSubmit={this.submit} />;
+  }
+}
 
-export const validEmail = mail => (mail.includes('@') ? undefined : 'User E-mail not valid');
-
-const passwordLongError = 'Password must be 8 characters long';
-export const passwordLength = password => (password.length < 8 ? undefined : passwordLongError);
+export default Login;
